WebApp Vs2.2

+++ ClassBackendBotanica/Artigo.cs
@@ -11,7 +11,7 @@ namespace ClassBackendBotanica
         public int Id { get; set; }
         
         public string Nome { get; set; }
-        public int Preco { get; set; }
+        public double Preco { get; set; }
         public int? CategoriaId { get; set; }
         public Categoria Categoria { get; set; }
